Deglob files to lint due to ESLint v5 requirements
As of ESLint v5, "Linting nonexistent files from the command line is now
a fatal error". Due to this, the file globs that are passed to ESLint
must first be deglobbed into an actual array of files to lint before
being passed.

More info here: https://eslint.org/docs/user-guide/migrating-to-5.0.0#-linting-nonexistent-files-from-the-command-line-is-now-a-fatal-error